🎨 Rename and add details in feature
